Remove initialize_component_exchange
Julia: "you likely won't need the `initialize_component_model_exchange` function you've defined.
We need this in our coupled ClimaAtmos/ClimaLand case because the two components have diagnostic variables that depend on each other and we need to set them correctly for restarting.
In your case, as long as the initial conditions are reasonable you won't need to include this."