Add barebones structure for restarting model state
Added a function restart_sims! to restart a simulation (could be part of Checkpointer.jl?).
As a simple test, restart_sims!() is called using the same callback frequency as checkpointing the model.
That is, everytime the model state is written, it is re-read from the same file.
This does not work at the moment since loading fields with a VF data layout is not supported by ClimaCore.