Remove CIT boundary mapping
It breaks my heart a bit but this feature introduces too many constraints (e.g., the whole solver being stuck to [0,dt_cpl]) and logic in the right-hand sides.
Implementing this properly should be part of the coupler, not the solvers as was the case now.
This way, the solver could also take care of properly adjusting the time coordinate for a given coupling window, and provide an API to evaluate the coupling fields at a given time, i.e., T_A(t) inside the right-hand side.